Build success! bags and style fixed
