+++ src/main/java/core/basesyntax/model/Dog.java
@@ -1,0 +1,4 @@
+package core.basesyntax.model;
+
+public class Dog {
+}
